move datahub plugin core into main codebase
